requirements.txt
pyjwt[crypto]
ollama
requests